ister: add --target-dir option
Allow users selecting the installation root directory (AKA target directory).
By default we just create a temporary directory, but if users have a
preference, they can select their own.

+++ ister.py
@@ -311,7 +311,25 @@ def create_filesystems(template):
                             raise_exception=False)
 
 
-def setup_mounts(template):
+def create_target_dir(args, template):
+    """Create the target root directory
+    """
+
+    if args.target_dir:
+        target_dir = args.target_dir
+        if not os.path.exists(target_dir):
+            raise Exception("Target directory {0} does not exist".format(target_dir))
+    else:
+        try:
+            prefix = "ister-" + str(template["Version"]) + "-"
+            target_dir = tempfile.mkdtemp(prefix=prefix)
+        except Exception:
+            raise Exception("Failed to setup mounts for install")
+
+    LOG.debug("Installation target directory: {0}".format(target_dir))
+    return target_dir
+
+def setup_mounts(target_dir, template):
     """Mount target folder
 
     Returns target folder name
@@ -319,12 +337,6 @@ def setup_mounts(template):
     This function will raise an Exception on finding an error.
     """
     LOG.info("Setting up mount points")
-    try:
-        prefix = "ister-" + str(template["Version"]) + "-"
-        target_dir = tempfile.mkdtemp(prefix=prefix)
-        LOG.debug("Using temporary directory: {0}".format(target_dir))
-    except Exception:
-        raise Exception("Failed to setup mounts for install")
 
     has_boot = False
 
@@ -813,8 +825,10 @@ def cleanup(args, template, target_dir, raise_exception=True):
             run_command("lsof {}/boot".format(target_dir),
                         raise_exception=raise_exception)
 
-        run_command("rm -fr {}".format(target_dir),
-                    raise_exception=raise_exception)
+        if not args.target_dir:
+            # --target-dir was not used.
+            run_command("rm -fr {}".format(target_dir),
+                        raise_exception=raise_exception)
 
     # Turn off any swap devices we enabled
     for fst in template["FilesystemTypes"]:
@@ -1474,7 +1488,8 @@ def install_os(args, template):
         if template["DestinationType"] == "virtual":
             map_loop_device(template)
         create_filesystems(template)
-        target_dir = setup_mounts(template)
+        target_dir = create_target_dir(args, template)
+        setup_mounts(target_dir, template)
         copy_os(args, template, target_dir)
         add_users(template, target_dir)
         set_hostname(template, target_dir)
@@ -1557,6 +1572,9 @@ def handle_options(sys_args):
                         help="Path to swupd state dir")
     group.add_argument("-F", "--fast-install", action="store_true",
                         help="Move swupd state dir inside image for a faster install")
+    parser.add_argument("--target-dir", action="store",
+                        default=None,
+                        help="Target root directory path, 'mktemp' by default")
     parser.add_argument("-d", "--dnf-config", action="store",
                         default=None,
                         help="DNF configuration file for installing packages")
